\name{PracticePackage2-package}
\alias{PracticePackage2-package}
\alias{PracticePackage2}
\docType{package}
\title{
\packageTitle{PracticePackage2}
}
\description{
\packageDescription{PracticePackage2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PracticePackage2}
\packageIndices{PracticePackage2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PracticePackage2}

Maintainer: \packageMaintainer{PracticePackage2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
